+++ src/keycode.h
@@ -97,6 +97,8 @@ public:
   OP(BACKWARD_WORD, "backward-word")               /* ALT-B / ALT-LEFT */                          \
   OP(FORWARD_WORD, "forward-word")                 /* ALT-F / ALT-RIGHT */                         \
   OP(NEWLINE, "newline")                           /* ALT-ENTER */                                 \
+  OP(YANK, "yank")                                 /* CTRL-Y */                                    \
+  OP(YANK_POP, "yank-pop")                         /* ALT-Y */                                     \
   OP(INSERT_KEYCODE, "insert-keycode")             /* CTRL-V */                                    \
   OP(BRACKET_PASTE, "bracket-paste")               /* ESC [200~ */                                 \
   OP(CUSTOM, "%custom")                            /* for custom action */
@@ -112,7 +114,8 @@ enum class EditActionType : unsigned char {
   OP(REPLACE_WHOLE_ACCEPT, "replace-whole-accept")                                                 \
   OP(REPLACE_LINE, "replace-line")                                                                 \
   OP(INSERT, "insert")                                                                             \
-  OP(HIST_SELCT, "hist-select")
+  OP(HIST_SELCT, "hist-select")                                                                    \
+  OP(KILL_RING_SELECT, "kill-ring-select")
 
 enum class CustomActionType : unsigned char {
 #define GEN_ENUM(E, S) E,